added providers back to Web3.providers
