+++ packages/web3-eth-abi/src/types/bool.js
