+++ packages/web3-eth-abi/src/types/bytes.js
