+++ packages/web3-eth-abi/src/types/dynamicbytes.js
