+++ packages/web3-eth-abi/src/types/string.js
