+++ packages/web3-eth-contract/src/index.js
@@ -37,7 +37,7 @@ var utils = require('web3-utils');
 var Subscription = require('web3-core-subscriptions').subscription;
 var formatters = require('web3-core-helpers').formatters;
 var promiEvent = require('web3-core-promiEvent');
-var coder = require('web3-eth-abi').solidity;
+var coder = require('web3-eth-abi');
 
 
 /**
