+++ packages/web3-eth/package.json
@@ -11,6 +11,7 @@
         "web3-core-subscriptions": "^1.0.0",
         "web3-core-method": "^1.0.0",
         "web3-eth-iban": "^1.0.0",
+        "web3-eth-abi": "^1.0.0",
         "web3-eth-contract": "^1.0.0",
         "web3-utils": "^1.0.0",
         "underscore": "^1.8.3"
