+++ test/abi.decodeParam-ethjs.js_
