+++ test/abi.decodeParam.js
@@ -1,6 +1,6 @@
 var chai = require('chai');
 var assert = chai.assert;
-var coder = require('../packages/web3-eth-abi').solidity;
+var coder = require('../packages/web3-eth-abi');
 
 // TODO check line 108 again!
