+++ test/abi.encodeParam.js
@@ -1,6 +1,6 @@
 var chai = require('chai');
 var assert = chai.assert;
-var coder = require('../packages/web3-eth-abi').solidity;
+var coder = require('../packages/web3-eth-abi');
 
 
 describe('lib/solidity/coder', function () {
